+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## master
 
+- Add Msgpack support (`--msgpack`). ([@palkan][])
+
+- Add WS sub-protocols support. ([@palkan][])
+
+- Upgrade to mruby 3.0. ([@palkan][])
+
 ## 0.3.1 (2020-06-30)
 
 - Add `--channel-params` option. ([@palkan][])